requirements.txt
python-terrier
sentence-transformers
fast-forward-indexes